Fix: Command 패턴으로 변경

+++ src/main/java/shoppingMall/controller/adminServlet/DeleteUserCommand.java
@@ -1,21 +1,26 @@
 package shoppingMall.controller.adminServlet;
 
 import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import shoppingMall.controller.Command;
 import shoppingMall.service.UserService;
 
 import java.io.IOException;
 
-@WebServlet("/admin/deleteMember")
-public class DeleteApproveServlet extends HttpServlet {
+public class DeleteUserCommand implements Command {
+
     private final UserService userService = new UserService();
+
     @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String userId = request.getParameter("userId");
-        userService.deleteUserById(userId);
+
+        if (userId != null && !userId.isEmpty()) {
+            userService.deleteUserById(userId); // userId로 유저 삭제
+        }
+
+        // 삭제 후 관리자 페이지로 리다이렉트
         response.sendRedirect(request.getContextPath() + "/adminPage.do");
     }
 }
